Scheduler Optimization: sched. is still in ticks, context switches are not inserted into the list and number of ns2tick and tick2ns conversions reduced

+++ miosix/kernel/kernel.h
@@ -519,7 +519,7 @@ public:
      * CANNOT be called when the kernel is paused.
      */
     static void sleep(unsigned int ms);
-    static void nanoSleep(unsigned int ns);
+    static void nanoSleep(long long ns);
     /**
      * Put the thread to sleep until the specified absolute time is reached.
      * If the time is in the past, returns immediately.
